Initial Commint for Talents
Started working on talents. Got it so it can read the raw data from the pdf and convert it into the objects. Just need to Save them into JSON (Which is super easy) then I won't be reading the raw text again (except if there's another massive dump of talents).